kubernetes-operators/mysql-operator/src/requirements.txt
kopf == 1.37.4
kubernetes == 31.0.0
PyYAML == 6.0.1
